requirements.txt
xmljson==0.2.0
fpdf==1.7.2
Pillow==5.4.1